Fixed some more ways
